improves requires, bumps version
