+++ lib/livetext/errors.rb
