+++ lib/livetext/formatline.rb
