+++ lib/livetext/functions.rb
