+++ lib/livetext/global_helpers.rb
