+++ lib/livetext/handler.rb
