+++ lib/livetext/handler/mixin.rb
