+++ lib/livetext/parser/general.rb
