+++ lib/livetext/processor.rb
