+++ lib/livetext/skeleton.rb
@@ -1,0 +1,17 @@
+require_relative 'parser/string'    # FIXME eh, should fix this
+
+# Class Livetext skeleton (top level).
+
+class Livetext
+  module Handler
+  end
+
+  module ParsingConstants
+  end
+
+  class FormatLine < StringParser
+    module FunCall
+    end
+  end
+end
+
