+++ lib/livetext/standard.rb
