+++ lib/livetext/version.rb
@@ -1,0 +1,6 @@
+
+# Defining VERSION
+
+class Livetext
+  VERSION = "0.9.25"
+end
