+++ utils/bump-version
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 
-system("vi lib/livetext.rb +4")
+system("vi lib/livetext/version.rb +/VER")
 
 require 'date'
 require "livetext"
@@ -12,32 +12,4 @@ system("git pull")
 gs = `git status 2>&1`
 
 puts gs
-
-exit
-
-if gs =~ /not staged for commit/
-  puts gs
-  puts "\n\n Possible bad git status. Continue? (y/n)"
-  resp = gets.chomp
-  case resp 
-    when "y"
-      system("git add -u")
-      print "Commit message = "
-      msg = gets.chomp
-      system("git commit -m '#{msg}' && git push")
-      puts "Now continuing..."
-    else         # assume "no"
-      puts "Exiting.\n "
-      exit
-  end
-else
-end
-
-# update = <<-EOF
-#   gem build livetext.gemspec && 
-#   gem push livetext-*.gem && 
-#   gem uninstall -x livetext && 
-#   gem install livetext
-# EOF
-# system(update)
